+++ src/eca/messenger.clj
@@ -7,6 +7,8 @@
 (defprotocol IMessenger
   (chat-content-received [this data])
   (chat-cleared [this params])
+  (chat-status-changed [this params])
+  (chat-deleted [this params])
   (rewrite-content-received [this data])
   (tool-server-updated [this params])
   (config-updated [this params])